Fix drawer sidebar
